feat(prejoin) show connection status in exported prejoin screen
