+++ react/features/prejoin/components/PrejoinApp.js
@@ -2,11 +2,13 @@
 
 import { AtlasKitThemeProvider } from '@atlaskit/theme';
 import React from 'react';
+import { batch } from 'react-redux';
 
 import { BaseApp } from '../../../features/base/app';
 import { setConfig } from '../../base/config';
 import { createPrejoinTracks } from '../../base/tracks';
-import { initPrejoin } from '../actions';
+import { getConferenceOptions } from '../../conference/functions';
+import { initPrejoin, makePrecallTest } from '../actions';
 
 import Prejoin from './Prejoin';
 
@@ -70,7 +72,10 @@ export default class PrejoinApp extends BaseApp<Props> {
 
             const tracks = await tryCreateLocalTracks;
 
-            dispatch(initPrejoin(tracks, errors));
+            batch(() => {
+                dispatch(initPrejoin(tracks, errors));
+                dispatch(makePrecallTest(getConferenceOptions(store.getState())));
+            });
         });
     }
